architecture refactoring: add services
