--- .teamcity/pom.xml
@@ -1,104 +1,0 @@
-<?xml version="1.0"?>
-<project>
-  <modelVersion>4.0.0</modelVersion>
-  <name>TidybeeHub Config DSL Script</name>
-  <groupId>TidybeeHub</groupId>
-  <artifactId>TidybeeHub_dsl</artifactId>
-  <version>1.0-SNAPSHOT</version>
-
-  <parent>
-    <groupId>org.jetbrains.teamcity</groupId>
-    <artifactId>configs-dsl-kotlin-parent</artifactId>
-    <version>1.0-SNAPSHOT</version>
-  </parent>
-
-  <repositories>
-    <repository>
-      <id>jetbrains-all</id>
-      <url>https://download.jetbrains.com/teamcity-repository</url>
-      <snapshots>
-        <enabled>true</enabled>
-      </snapshots>
-    </repository>
-    <repository>
-      <id>teamcity-server</id>
-      <url>http://aaaaaaaaaaaaaa.freeboxos.fr:8111/app/dsl-plugins-repository</url>
-      <snapshots>
-        <enabled>true</enabled>
-      </snapshots>
-    </repository>
-  </repositories>
-
-  <pluginRepositories>
-    <pluginRepository>
-      <id>JetBrains</id>
-      <url>https://download.jetbrains.com/teamcity-repository</url>
-    </pluginRepository>
-  </pluginRepositories>
-
-  <build>
-    <sourceDirectory>${basedir}</sourceDirectory>
-    <plugins>
-      <plugin>
-        <artifactId>kotlin-maven-plugin</artifactId>
-        <groupId>org.jetbrains.kotlin</groupId>
-        <version>${kotlin.version}</version>
-
-        <configuration/>
-        <executions>
-          <execution>
-            <id>compile</id>
-            <phase>process-sources</phase>
-            <goals>
-              <goal>compile</goal>
-            </goals>
-          </execution>
-          <execution>
-            <id>test-compile</id>
-            <phase>process-test-sources</phase>
-            <goals>
-              <goal>test-compile</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>org.jetbrains.teamcity</groupId>
-        <artifactId>teamcity-configs-maven-plugin</artifactId>
-        <version>${teamcity.dsl.version}</version>
-        <configuration>
-          <format>kotlin</format>
-          <dstDir>target/generated-configs</dstDir>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.jetbrains.teamcity</groupId>
-      <artifactId>configs-dsl-kotlin-latest</artifactId>
-      <version>${teamcity.dsl.version}</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jetbrains.teamcity</groupId>
-      <artifactId>configs-dsl-kotlin-plugins-latest</artifactId>
-      <version>1.0-SNAPSHOT</version>
-      <type>pom</type>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jetbrains.kotlin</groupId>
-      <artifactId>kotlin-stdlib-jdk8</artifactId>
-      <version>${kotlin.version}</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jetbrains.kotlin</groupId>
-      <artifactId>kotlin-script-runtime</artifactId>
-      <version>${kotlin.version}</version>
-      <scope>compile</scope>
-    </dependency>
-  </dependencies>
-</project>
