+++ DataProcessing/Controllers/DataProcessingController.cs
@@ -1,0 +1,15 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace DataProcessingService.Controllers
+{
+    [ApiController]
+    [Route("[controller]")]
+    public class DataProcessingController : ControllerBase
+    {
+        [HttpGet]
+        public IActionResult Get()
+        {
+            return Ok("DataProcessing Service hit");
+        }
+    }
+}
