+++ MicroservicesPoC.sln
@@ -1,0 +1,65 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34221.43
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MicroservicesPoC", "MicroservicesPoC.vcxproj", "{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApiGateway", "ApiGateway\ApiGateway.csproj", "{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Auth", "Auth\Auth.csproj", "{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Debug|Any CPU.Build.0 = Debug|x64
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Debug|x64.ActiveCfg = Debug|x64
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Debug|x64.Build.0 = Debug|x64
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Debug|x86.ActiveCfg = Debug|Win32
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Debug|x86.Build.0 = Debug|Win32
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Release|Any CPU.ActiveCfg = Release|x64
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Release|Any CPU.Build.0 = Release|x64
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Release|x64.ActiveCfg = Release|x64
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Release|x64.Build.0 = Release|x64
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Release|x86.ActiveCfg = Release|Win32
+		{5554CF9B-BFFD-4A6D-A37D-357A0DE95AC9}.Release|x86.Build.0 = Release|Win32
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Debug|x64.Build.0 = Debug|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Debug|x86.Build.0 = Debug|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Release|x64.ActiveCfg = Release|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Release|x64.Build.0 = Release|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Release|x86.ActiveCfg = Release|Any CPU
+		{DAFC5295-9155-4D3D-B5DC-C74232A1EE71}.Release|x86.Build.0 = Release|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Debug|x64.Build.0 = Debug|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Debug|x86.Build.0 = Debug|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Release|x64.ActiveCfg = Release|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Release|x64.Build.0 = Release|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Release|x86.ActiveCfg = Release|Any CPU
+		{9BBB10EE-F7AF-48F1-B411-0A0C67BD867C}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {72B3D291-4C6F-45FE-BB95-7F9DD9F8FF55}
+	EndGlobalSection
+EndGlobal
