--- Models/AgentMetadataModel.cs
@@ -1,12 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace api.Models
-{
-
-    public class AgentMetadataModel
-    {
-        [Key]
-        public int Id { get; set; }
-        public string? Json { get; set; }
-    }
-}
