--- Models/AgentStatusModel.cs
@@ -1,9 +1,0 @@
-namespace api.Models {
-
-    public enum AgentStatusModel
-    {
-        TroubleShooting = -1,
-        Disconnected = 0,
-        Connected = 1
-    }
-}
