--- Program.cs
@@ -1,54 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using tidybee_hub.Context;
-using tidybee_hub.Repository;
-
-var builder = WebApplication.CreateBuilder(args);
-
-builder.Services.AddHttpClient();
-builder.Services.AddControllers();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-builder.Services.AddDbContext<DatabaseContext>(options =>
-    options.UseSqlite(builder.Configuration.GetConnectionString("DatabaseConnection")));
-builder.Services.AddScoped<AgentRepository>();
-
-var app = builder.Build();
-
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-if (app.Configuration.GetValue<bool>("EnableAutoMigration"))
-{
-    using var scope = app.Services.CreateScope();
-    var services = scope.ServiceProvider;
-
-    try
-    {
-        var dbContext = services.GetRequiredService<DatabaseContext>();
-        dbContext.Database.EnsureCreated();
-        dbContext.Database.Migrate();
-    }
-    catch (Exception ex)
-    {
-        var logger = services.GetRequiredService<ILogger<Program>>();
-        logger.LogError(ex, "An error occurred while migrating the database.");
-    }
-}
-
-app.UseMiddleware<ProxyMiddleware>();
-
-app.UseHttpsRedirection();
-
-app.UseAuthorization();
-
-app.UseCors(policy => policy
-    .AllowAnyOrigin()
-       .AllowAnyMethod()
-          .AllowAnyHeader());
-
-app.MapControllers();
-
-app.Run();
