--- appsettings.Development.json
@@ -1,23 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "Kestrel": {
-    "Endpoints": {
-      "Http": {
-        "Url": "http://localhost:5000"
-      },
-      "Https": {
-        "Url": "https://localhost:5001"
-      }
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-      "DatabaseConnection": "Data Source=DATABASE_NAME.db"
-  },
-  "EnableAutoMigration": true
-}
