--- tidybee-hub.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34024.191
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "tidybee-hub", "tidybee-hub.csproj", "{E48C0214-4146-46F1-AE70-E5810D507D02}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E48C0214-4146-46F1-AE70-E5810D507D02}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E48C0214-4146-46F1-AE70-E5810D507D02}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E48C0214-4146-46F1-AE70-E5810D507D02}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E48C0214-4146-46F1-AE70-E5810D507D02}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D637303A-11CF-4FAC-BFF8-5F595F3DD290}
-	EndGlobalSection
-EndGlobal
